Rename "NewCreateBatchChangePage" to "CreateOrEditBatchChangePage"
