+++ client/web/src/enterprise/user/routes.tsx
@@ -5,7 +5,7 @@ import { userAreaRoutes } from '../../user/area/routes'
 import { UserAreaRoute, UserAreaRouteContext } from '../../user/area/UserArea'
 import { lazyComponent } from '../../util/lazyComponent'
 import { CreateBatchChangePageProps } from '../batches/create/CreateBatchChangePage'
-import { NewCreateBatchChangePageProps } from '../batches/create/NewCreateBatchChangePage'
+import { CreateOrEditBatchChangePageProps } from '../batches/create/CreateOrEditBatchChangePage'
 import { NamespaceBatchChangesAreaProps } from '../batches/global/GlobalBatchChangesArea'
 import { SHOW_BUSINESS_FEATURES } from '../dotcom/productSubscriptions/features'
 import { enterpriseNamespaceAreaRoutes } from '../namespaces/routes'
@@ -15,9 +15,9 @@ const NamespaceBatchChangesArea = lazyComponent<NamespaceBatchChangesAreaProps, 
     'NamespaceBatchChangesArea'
 )
 
-const NewCreateBatchChangePage = lazyComponent<NewCreateBatchChangePageProps, 'NewCreateBatchChangePage'>(
-    () => import('../batches/create/NewCreateBatchChangePage'),
-    'NewCreateBatchChangePage'
+const CreateOrEditBatchChangePage = lazyComponent<CreateOrEditBatchChangePageProps, 'CreateOrEditBatchChangePage'>(
+    () => import('../batches/create/CreateOrEditBatchChangePage'),
+    'CreateOrEditBatchChangePage'
 )
 
 const CreateBatchChangePage = lazyComponent<CreateBatchChangePageProps, 'CreateBatchChangePage'>(
@@ -50,7 +50,7 @@ export const enterpriseUserAreaRoutes: readonly UserAreaRoute[] = [
     {
         path: '/batch-changes/:batchChangeName/edit',
         render: ({ match, ...props }: UserAreaRouteContext & RouteComponentProps<{ batchChangeName: string }>) => (
-            <NewCreateBatchChangePage
+            <CreateOrEditBatchChangePage
                 {...props}
                 initialNamespaceID={props.user.id}
                 batchChangeName={match.params.batchChangeName}
